backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.19.0